% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quiz1.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number indicating the state}

\item{year}{The input year}
}
\value{
A plot showing the latitude and longitude points where the accident happened
}
\description{
This function takes a state number and year as the input, and then it takes the state name
corresponding to the state number. Then for all data points with NA, it ensures that no values
with NA are present. Finally, it makes a map of the latitudes and longitudes for that state.
Also, the function has error checks to ensure that the state number is valid, and also returns
NULL if there are no accidents in that state in that year
}
\examples{
\dontrun{fars_map_state(10, 2013)}

}

