% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quiz1.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file, whose data is to be extracted}
}
\value{
This function returns a data frame having data from the file
}
\description{
This function reads the data in the file specified by the file name, without the messages.
If the file does not exist, then the function stops the execution and returns an error
}
\examples{
\dontrun{data1 <- fars_read('accident_2013.csv')}

}

