% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quiz1.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years input for all years for which files are available}
}
\value{
The month and year from the file
}
\description{
This function takes all the years, and for each year, it obtains the filename for that year, then
reads that file, adds the year data, and then selects the months and year columns
}
\examples{
\dontrun{fars_read_years(c(2013, 2014))}

}

