% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quiz1.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years input for all years for which files are available}
}
\value{
The summarized data having number of entries for each month in each year
}
\description{
This function takes the year and month out of the data, and obtains the number of observations in
each month of each year
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014))}

}

