% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quiz1.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The input year}
}
\value{
This function prints the filename corresponding to the year requested
}
\description{
Since there is data available for each year, this function takes the input year and then prints the
filename corresponding to that year, and also makes the filename for that year.
}
\examples{
\dontrun{make_filename(2013)}

}

